agentspace-app/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
pydantic==1.10.7
python-multipart==0.0.6
google-cloud-storage==2.9.0
google-auth==2.17.3
pydub==0.25.1
aiofiles==23.1.0
jinja2==3.1.2
python-dotenv==1.0.0
